Use a single course-grained RwLock in the root
Discard the fine-grained locking as the overhead of locking in each node
is too high. Instead maintain a single RwLock in the root and take read
locks for almost all operations except for the ones that require us to
add a new node.

This is still about 20% slower than with no synchronization for
single-threaded benchmarks.