Add a _goban_ benchmark with a high cardinality
Complement the current benchmark suite with a _goban_ benchmark that has
both (a) high cardinality, and (b) expensive evaluation.

This benchmark show a different picture compared to the _sticks_
benchmark in that it does not hit very high lock congestion, so it
scales pretty well with the number of threads.