Improve performance of `try_expand` during high contention
Use a `binary_search_by_key` followed by an `insert` instead of a `push`
and `sort_unstable_by_key` when adding a new child as the later (a) does
not disallow duplicates, and (b) is O(n log n) instead of O(n).

This seems to improve the performance of `sticks/4` and `sticks/8` but
slightly reduce performance for low parallelism use-cases for some
reason.